+++ pkg/controller/deployment/recreate.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 
 	"k8s.io/kubernetes/pkg/apis/extensions"
-	unversionedclient "k8s.io/kubernetes/pkg/client/unversioned"
+	"k8s.io/kubernetes/pkg/client/retry"
 	"k8s.io/kubernetes/pkg/controller"
 	"k8s.io/kubernetes/pkg/util/wait"
 )
@@ -107,7 +107,7 @@ func (dc *DeploymentController) waitForInactiveReplicaSets(oldRSs []*extensions.
 		specReplicas := rs.Spec.Replicas
 		statusReplicas := rs.Status.Replicas
 
-		if err := wait.ExponentialBackoff(unversionedclient.DefaultRetry, func() (bool, error) {
+		if err := wait.ExponentialBackoff(retry.DefaultRetry, func() (bool, error) {
 			replicaSet, err := dc.rsLister.ReplicaSets(rs.Namespace).Get(rs.Name)
 			if err != nil {
 				return false, err
