+++ pkg/registry/core/service/portallocator/controller/repair.go
@@ -22,7 +22,7 @@ import (
 
 	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/errors"
-	client "k8s.io/kubernetes/pkg/client/unversioned"
+	"k8s.io/kubernetes/pkg/client/retry"
 	"k8s.io/kubernetes/pkg/registry/core/rangeallocation"
 	"k8s.io/kubernetes/pkg/registry/core/service"
 	"k8s.io/kubernetes/pkg/registry/core/service/portallocator"
@@ -61,7 +61,7 @@ func (c *Repair) RunUntil(ch chan struct{}) {
 
 // RunOnce verifies the state of the port allocations and returns an error if an unrecoverable problem occurs.
 func (c *Repair) RunOnce() error {
-	return client.RetryOnConflict(client.DefaultBackoff, c.runOnce)
+	return retry.RetryOnConflict(retry.DefaultBackoff, c.runOnce)
 }
 
 // runOnce verifies the state of the port allocations and returns an error if an unrecoverable problem occurs.
